requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-mathfilters==1.0.0
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
mccabe==0.7.0
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sorl-thumbnail==12.9.0
sqlparse==0.4.3
stripe==5.2.0
urllib3==1.26.14
